requirements.txt
Django==4.2.7
djangorestframework==3.14.0
markdown==3.5.1
pillow==10.1.0
docxtpl==0.16.7
